\name{DES2}
\alias{DES2}
\docType{data}
\title{DES2 \{dispersion-package\}}
\description{This is an .rda file containing a design in a matrix.}
\usage{data(DES2)}
\format{Three columns and eighteen rows.}
\details{This design is presented in Table 2 of [1], referred to as design II, obtained by optimizing the $(DP)_S$ criterion. It allows the maximum degree of freedom to estimate pure error. It is a completely randomized design for three factors in the spherical region.}
\source{[1] GILMOUR, S. G.; TRINCA, L. A. ?Optimum design of experiments for statistical inference (with discussion). J. Roy. Statist. Soc., Series C (Applied Statistics), 61,345-401. 2012.}
\keyword{datasets}
