\name{DES4}
\alias{DES4}
\docType{data}
\title{DES4 \{dispersion-package\}}
\description{This is an .rda file containing a design in a matrix.}
\usage{data(DES4)}
\format{Five columns and forty-two rows.}
\details{This design is presented in Table 10 of [1], obtained by optimizing the D criterion, with five factors (one hard-to-set and four easy-to-set) in 21 whole plots of two sub-plots each, in a cuboidal region.}
\source{[1] GOOS, P.; JONES, B. I-optimal versus D-optimal split-plot response surface designs.
Journal of Quality Technology. v. 44, n. 2, p. 85-101, 2012.}
\keyword{datasets}
