\name{dispersion-package}
\alias{dispersion}
\title{Dispersion statistics for small samples.}
\docType{package}

\description{
[add a description here!]
}


\author{
Paul Kattuman, Thilo Klein, Jun Ma
}


\references{
Kattuman, P.A. and Ma, J. (2014). Dispersion Statistics for
Small Samples. \emph{Working Paper}, Cambridge Judge Business School.
}


\section{How to cite this package}{
  Whenever using this package, please cite as

  Kattuman, P.A., Klein, T. and Ma, J. (2014). dispersion: Dispersion statistics for small samples. R package version 0.1.0.
}


\seealso{
  \href{http://cran.r-project.org/web/packages/ineq/index.html}{ineq}, 
  \href{http://cran.r-project.org/web/packages/reldist/index.html}{reldist}
}


\keyword{package}