% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispunif3}
\alias{dispunif3}
\title{Dispersion statistics for small samples}
\usage{
dispunif3(s = c(1, 3, 7), lb = 0.1, ub = 0.9, bss = 2000)
}
\arguments{
\item{s}{The observed sample; (not in share vector form) read in from an external file}

\item{lb}{lower bound for confidence interval}

\item{ub}{upper bound for confidence interval}

\item{bss}{number of bootstrap samples}
}
\value{
\code{dispunif3} returns a matrix with two columns for dispersion measures
P(s >= S) and P(s <= S).
}
\description{
This program generates dispersion measures P(S more equal than s) and P(S less
equal than s) for data vectors of size n. The approach is to resample from the
data vector, choosing 3 elements randomly, and find the defined measures
with S drawn from the uniform distribution on the unit 2-simplex, with bss resamples.
}
\examples{
## share vector of length 3
dispunif3(s=c(1,3,7),lb=0.1,ub=0.9,bss=2000)

## share vector of length larger 3
dispunif3(s=c(1,2,3,4,5,6,7,8,9,10),lb=0.1,ub=0.9,bss=2000)
}
\author{
Paul Kattuman, Thilo Klein, Jun Ma
}
\references{
Kattuman, P.A. and Ma, J. (2014). Dispersion Statistics for
Small Samples. \emph{Working Paper}, Cambridge Judge Business School.
}
\keyword{dispersion}
\keyword{statistics}

