\name{libor}
\docType{data}
\alias{libor}

\title{Libor dataset}

\description{
[add a description here!]
}

\usage{data(libor)}

\format{
This data frame contains the following columns:
\describe{
\item{Barclays}{.}
\item{UBS}{.}
\item{BTMU}{.}
\item{BOA}{.}
\item{CITI}{.}
\item{CRSW}{.}
\item{DEUT}{.}
\item{HBOSSocGen}{.}
\item{HSBC}{.}
\item{JPM}{.}
\item{LLOYDS}{.}
\item{NORIN}{.}
\item{RABO}{.}
\item{AbbRBC}{.}
\item{RBS}{.}
\item{WestLB}{.}
}}

\source{
[add a source here!]
}

\references{
[add a reference here!]
}

\keyword{datasets}

