% Generated by roxygen2 (4.0.1): do not edit by hand
\name{robustmeasure}
\alias{robustmeasure}
\title{Dispersion statistics for small samples (general)}
\usage{
robustmeasure(s, shape, l1, n, bss, lb, ub)
}
\arguments{
\item{shape}{Gamma distribution parameter for shape, for D(1), shape=1: uniform distribution on the simplex}

\item{s}{The observed sample (not in share vector form).}

\item{n}{bootstrap sample size , typically 3}

\item{bss}{number of bootstrap samples}

\item{lb}{lower bound 5 or .05 or 10 or .1}

\item{ub}{upper bound 95 or .95 or 90 or .9}

\item{l1}{the number of samples drawn for prior - S~D(lambda,) to calculate by simulation}
}
\value{
\code{robustmeasure} returns samp dists of different measures for different alphas and
power test proportions at the right of the right side cutoff and left of the left side
cutoff  of the samp dist for different measures.
}
\description{
This program generates sampling distributions of CoV, Gini, Entropy(H),
GE, HHI, and P(S more equal than s), P(S less equal than s) from draws from Pareto
distribution, for various alpha values.
Also generates data for the power curve for all the indices CoV, Gini, Entropy(H), GE,
HHI and P(S more equal than s), P(S less equal than s) for the left end of the samp dist
for chosen size of the test: ie. prop from the true dist to the left of the 5# (1#) cutoff
from the null hypothesized  distribution for the right end of the samp dist for chosen
size of the test: ie. prop from the true dist to the right of the 95# (99#) cutoff from
the null hypothesized dist.
}
\examples{
## share vector of length 3
robustmeasure(shape=1,s=c(1,3,7),n=3,bss=2000,lb=0.1,ub=0.9,l1=1000)

## share vector of length larger 3
robustmeasure(shape=1,s=c(1,2,3,4,5,6,7,8,9,10),n=3,bss=2000,lb=0.1,ub=0.9,l1=100)

\dontrun{
## matrix of share vectors
data(libor)
robustmeasure(s=libor, shape=1, l1=100, n=3, bss=2000, lb=10, ub=90)
}
}
\author{
Paul Kattuman, Thilo Klein, Jun Ma
}
\references{
Kattuman, P.A. and Ma, J. (2014). Dispersion Statistics for
Small Samples. \emph{Working Paper}, Cambridge Judge Business School.
}
\keyword{dispersion}
\keyword{statistics}

